Debugging
Modified the first turning case in OP 4 so that once the encoder
difference is greater than the set amount, the while loop will escape
and the TimeTurning variable will be increased. This should circumvent
the case where the difference was enough to get out of the loop but not
reach the conditional statement (leaving the robot turning).